Added pictures for name, measurements and the new preview pictures for ChosenWall-xml.
Also added so the coresponding pictures connects to their imageviews for a new and better layout.

Also made a litle change in the switch as it were done incorrectly. it worked but it were connected to the wrong pictures.

lastly i've changed some picture codes layout to be in the same order as the catalog.